Visualizer.java: The previous version had 3 pipes per channel but this was a waste. I changed it to 2 pipes per channel.     Added a constant for the minimum size of the radius     Moved the initialization of pipesToUse into the constructor     Added pitch bends into the Visualizer.java with a method to handle it     Considering flashes have been added to the Beat.java, I had to reset the light emissions after the beats are drawn     Added a method to process pitch changes
TonalProcessor.java: To correspond with changing the pipes per channel from 3 to 2. I had to change the intervals to: 0 >= z <= 41 and 42 <= z <= 127

threads/ThreadPreprocessor.java: When the preprocessor is started, I pause the FPSAnimator.java in JOGL then resume() it when the preprocessor is done. I changed how the FPSAnimator.java is accessed
    by making it thread safe in gui/GUI.java therefore I changed the calls in this class to correspond to the changes in the gui.

Pipe.java: Added static constants to keep track of the Indicies that the vertex, colour and normals are stored.
    Added a gluSpehre at the beginning of the pipe therefore I had to keep track of the current coordinates and the current radius
    Have corrected the normals so the lighting works now
    ****Added some code in the draw method where it will not animate the pipe if the pipe has nothing to draw. This was  BIG CHANGE AND INCREASES performance
    Made sure that the constants are used in the code to make the class as dynamic as possible
    Added many comments

MidiNoteReceiver.java: Earlier I changed how often the GUI was updated, there was something wrong with this. The problem was when the slider
    made to go back before the current time and then the slider would not update, therefore I fixed this.
    I added in the ability to send pitch changes to the visualizer/Visualizer.java

Gui.java: Added the ability to to cap the FPS. This was needed after the pipe.java optimization mentioned above.
    Changed the way the animator is accessed. THis time it must be accessed through a thread.

Beat.java: Added in flashing