This revision fixes a major issue that has plagued the program for 2 years. The main GUI that displays the MidiPlayer controls was created 2 years ago in the first MusicViz that interfaced with Flash. This section has carried forward for every other music visualization.
The issue is that the program uses a ridiculously high amount of CPU. Through testing and rewriting in  a test project, the artchitecture of the program did not cause the CPU usage spike. Looking closely at the code, the CPU spike happens only after a song is in the playlist GUI. Therefore the JList is most likley the culprit. To check this, print statements were added into all the methods that are invoked whenever the JList is drawn or changed through an event.

All listeners functioned correctly; however, the JList also has a FileCellRenderer.java, which describes how to display the list. Through print statements, it was found that this class was constantly being called (endlessly) and therefore causing the CPU spike.

All events are handled by the EventDispatch Thread. When a GUI item is clicked the EventDispatch thread handles it. Therefore whenever a song is added to the JList, its listeners are called as well as the FileCellRenderer. The problem was that the fileCell renderer had the following call: list.repaint(). This caused the EventDispatch thread to call itself again.

Hence whenever the JList is changed, the file cell renderer is called and then the file cell renderer puts in a request to redraw its self. Therefore it is an infinite loop that never ends causing a CPU spike. The other issue is that the EventDispatch Thread also times the Visualization. Therefore the Visualizer is called endless causing another CPU spike and to counteract this, the VIsualizer creates noOps in order to slow itsself down. Therefore the list.repaint() is removed from the FileCellRenderer.java to solve this issue.

Also to stop the EventDispatch thread from becoming overrun, the ConcurrentVisualizer.java was made into a thread that handled queuing animations and the Visualizer (which displays the visualization) only draws the animations. Therefore this takes off all strain on the EventDispatch Thread.