This version is the demo version of the program.
There are a few minor changes and one set of semi-major changes.
The minor changes are:

1. The ConcurrentVisualizer.java's run() method did not poll all the channels. For each run. the thread polls all channels 10 times before sleeping.
The indicies were messed up and the method polled 10 channels 16 times. This caused some of the pipes to not display. This is now fixed. Also the method
has been cleaned up by removing If( queue.peek() != null ) and adding if( message == null ){ continue; }. This is much cleaner because it removes one level
of indenting.

2. Another small issue was that the InstrumentReceiver.java did not perform proper channel calculations. To calculate the channel of a note off event, take the status
of the note off and minus 128. This is because note off statuses range from 128 to 144. The part where the calculation was incorrect was the note on calculation.
To calculate the channel, it is the note on's status minus 144. It was incorrectly 128 before. This is fixed

3. When loading in a new song, the ConcurrentVisualizer.java's pipes are reset. This means that the animation variables are reset to initial values. This meant that
lastAlpha, lastFace, currentRadius, and the animation queues must be reset. In the initial attempt to make the reset operation thread safe, the animation queues
were changed to ConcurrentLinkedQueues. However, the lastAlpha, lastFace and currentRadius were not changed to Atomic Variables. This meant there was a chance of
odd read/rewrite operations. Therefore, these three variables were changed to AtomicReferences. These variables make sure of no thread blocking and do not store 
caches of variables. The accesses happen in real-time in the main memory and all writes to variables happen in real time.