This is more clean up and is demo material.
There was one more problem with the Midi program. The problem was that the sound would cut-out abruptly and never come back.
I was not able to find the problem directly, however, I believe I solved the problem indirectly. When getting the sequencer
that is already connected, a connection is made to the output. In this case the sound worked but after loading a few songs
the sound would cut out. Therefore I found manually getting the synthesizer and loading a soundbank-deluxe.gm fixed the problem.

I am not sure why though.

Considering a synthesizer was used directly, I am able to power all sounds off and turn the volume on and off more easily.