There are 6 threads created to assist in processing the messages sent through, such that the EventDispatch Thread and the Java Sound Thread are not calculating too much. These threads were not set as Daemon threads. A Daemon thread is no different than a User thread; however, daemon threads can be shut down with no exception thrown when all other threads are destroyed.
Daemon threads exist to serve user threads. Therefore if the user threads do not exist, there is no need to have the daemon thread.
Daemon threads will not cause the garbage disposer to hang when exiting. This might be the case in Java 1.7 or it could be
release candidate 8 of JOGL 2.0 causing the issues. 

However, if it does happen, an exception in 2Diposer is not really an issue when closing the programming. The JVM cleans everything up.
http://stackoverflow.com/questions/7769885/interruptedexception-after-cancel-file-open-dialog-1-6-0-26

Another issue appeared with Java 1.7. This issue is that the program will not load the external soundbank-deluxe.gm library causing the program
to close. However, the program still works and has sound when the soundbank-deluxe.gm does not load. This is something that has only surfaced
with Java 1,7. Therefore I changed the program to just display a notification and continue loading.